refactor: move tier methods into own directory
