+++ src/node-lib/curriculum-api-2023/queries/unitListing/unitListing.schema.ts
@@ -1,29 +1,10 @@
-import { programmeFieldsSchema } from "@oaknational/oak-curriculum-schema";
 import { z } from "zod";
 
-export const rawTierResponseSchema = z.array(
-  z.object({
-    programme_fields: programmeFieldsSchema,
-    programme_slug: z.string(),
-  }),
-);
+import { tierSchema } from "./tiers/tiers.schema";
 
 const aggregateSchema = z.object({
   count: z.number(),
 });
-
-export const tierCounts = z.object({
-  lessonCount: z.object({
-    aggregate: aggregateSchema,
-    nodes: z.array(z.object({ programme_fields: z.string() })),
-  }),
-  unitCount: z.object({
-    aggregate: aggregateSchema,
-    nodes: z.array(z.object({ programme_fields: z.string() })),
-  }),
-});
-
-export type TierCounts = z.infer<typeof tierCounts>;
 
 export const lessonCounts = z.object({
   lessonCount: z.object({
@@ -97,16 +78,6 @@ export type UnitData = z.infer<typeof unitData>;
 export const unitSchema = z.array(z.array(unitData).min(1));
 export type UnitsForProgramme = z.infer<typeof unitSchema>;
 
-export const tierSchema = z.array(
-  z.object({
-    tierSlug: z.string(),
-    tierTitle: z.string(),
-    tierProgrammeSlug: z.string(),
-    unitCount: z.number().nullish(),
-    lessonCount: z.number().nullish(),
-    tierOrder: z.number().nullable(),
-  }),
-);
 const unitListingData = z.object({
   programmeSlug: z.string(),
   keyStageSlug: z.string(),
@@ -132,8 +103,6 @@ const unitListingData = z.object({
 });
 
 export type UnitListingData = z.infer<typeof unitListingData>;
-
-export type TierSchema = z.infer<typeof tierSchema>;
 
 const unitListingSchema = z.object({
   programmeSlug: z.string(),
